+++ test/src/terminal_view_test.dart
@@ -1458,10 +1458,22 @@ void main() {
       await secondaryGesture.up();
       await tester.pump();
 
-      expect(output.where((value) => value.endsWith('M')), hasLength(2));
-      expect(output.where((value) => value.endsWith('m')), hasLength(2));
+      final backGesture = await tester.createGesture(
+        pointer: 3,
+        kind: PointerDeviceKind.mouse,
+        buttons: kBackMouseButton,
+      );
+      await backGesture.down(start);
+      await backGesture.moveTo(end);
+      await tester.pump();
+      await backGesture.up();
+      await tester.pump();
+
+      expect(output.where((value) => value.endsWith('M')), hasLength(3));
+      expect(output.where((value) => value.endsWith('m')), hasLength(3));
       expect(output, contains(startsWith('\x1b[<0;')));
       expect(output, contains(startsWith('\x1b[<2;')));
+      expect(output, contains(startsWith('\x1b[<128;')));
       expect(controller.selection, isNull);
 
       controller.dispose();